merge upstream

+++ src/modules/mavlink/mavlink_log_handler.cpp
@@ -39,7 +39,7 @@
 #include <sys/stat.h>
 #include <time.h>
 
-#define MOUNTPOINT PX4_ROOTFSDIR "/fs/microsd"
+#define MOUNTPOINT PX4_STORAGEDIR
 
 static const char *kLogRoot    = MOUNTPOINT "/log";
 static const char *kLogData    = MOUNTPOINT "/logdata.txt";
@@ -429,6 +429,7 @@ LogListHelper::get_log_data(uint8_t len, uint8_t *buffer)
 
 	if (offset && fseek(current_log_filep, offset, SEEK_CUR)) {
 		fclose(current_log_filep);
+		current_log_filep = nullptr;
 		PX4LOG_WARN("MavlinkLogHandler::get_log_data Seek error in %s\n", current_log_filename);
 		return 0;
 	}
@@ -475,10 +476,13 @@ LogListHelper::_init()
 		if (result->d_type == PX4LOG_DIRECTORY) {
 			time_t tt = 0;
 			char log_path[128];
-			snprintf(log_path, sizeof(log_path), "%s/%s", kLogRoot, result->d_name);
+			int ret = snprintf(log_path, sizeof(log_path), "%s/%s", kLogRoot, result->d_name);
+			bool path_is_ok = (ret > 0) && (ret < (int)sizeof(log_path));
 
-			if (_get_session_date(log_path, result->d_name, tt)) {
-				_scan_logs(f, log_path, tt);
+			if (path_is_ok) {
+				if (_get_session_date(log_path, result->d_name, tt)) {
+					_scan_logs(f, log_path, tt);
+				}
 			}
 		}
 	}
@@ -534,12 +538,15 @@ LogListHelper::_scan_logs(FILE *f, const char *dir, time_t &date)
 				time_t  ldate = date;
 				uint32_t size = 0;
 				char log_file_path[128];
-				snprintf(log_file_path, sizeof(log_file_path), "%s/%s", dir, result->d_name);
+				int ret = snprintf(log_file_path, sizeof(log_file_path), "%s/%s", dir, result->d_name);
+				bool path_is_ok = (ret > 0) && (ret < (int)sizeof(log_file_path));
 
-				if (_get_log_time_size(log_file_path, result->d_name, ldate, size)) {
-					//-- Write result->out to list file
-					fprintf(f, "%u %u %s\n", (unsigned)ldate, (unsigned)size, log_file_path);
-					log_count++;
+				if (path_is_ok) {
+					if (_get_log_time_size(log_file_path, result->d_name, ldate, size)) {
+						//-- Write result->out to list file
+						fprintf(f, "%u %u %s\n", (unsigned)ldate, (unsigned)size, log_file_path);
+						log_count++;
+					}
 				}
 			}
 		}
@@ -601,20 +608,27 @@ LogListHelper::delete_all(const char *dir)
 
 		if (result->d_type == PX4LOG_DIRECTORY && result->d_name[0] != '.') {
 			char log_path[128];
-			snprintf(log_path, sizeof(log_path), "%s/%s", dir, result->d_name);
-			LogListHelper::delete_all(log_path);
+			int ret = snprintf(log_path, sizeof(log_path), "%s/%s", dir, result->d_name);
+			bool path_is_ok = (ret > 0) && (ret < (int)sizeof(log_path));
 
-			if (rmdir(log_path)) {
-				PX4LOG_WARN("MavlinkLogHandler::delete_all Error removing %s\n", log_path);
+			if (path_is_ok) {
+				LogListHelper::delete_all(log_path);
+
+				if (rmdir(log_path)) {
+					PX4LOG_WARN("MavlinkLogHandler::delete_all Error removing %s\n", log_path);
+				}
 			}
 		}
 
 		if (result->d_type == PX4LOG_REGULAR_FILE) {
 			char log_path[128];
-			snprintf(log_path, sizeof(log_path), "%s/%s", dir, result->d_name);
+			int ret = snprintf(log_path, sizeof(log_path), "%s/%s", dir, result->d_name);
+			bool path_is_ok = (ret > 0) && (ret < (int)sizeof(log_path));
 
-			if (unlink(log_path)) {
-				PX4LOG_WARN("MavlinkLogHandler::delete_all Error deleting %s\n", log_path);
+			if (path_is_ok) {
+				if (unlink(log_path)) {
+					PX4LOG_WARN("MavlinkLogHandler::delete_all Error deleting %s\n", log_path);
+				}
 			}
 		}
 	}
